corgi: the accumulator is u64, so a mask can be a byte
Answering one question settled both halves of it.

Reduce/Scan take ANY leaf width and accumulate at u64. The accumulator is u64
whatever the elements are, because a reduction of a narrow column routinely
exceeds it -- counting a byte mask is the motivating case, and it is what makes
the second half affordable. The widen is the leaf's UNSIGNED reading, so a narrow
signed or float column must be `cast` first; that caveat is the one crate::hash
already carries, for the same reason. Width and monoid both dispatch above the
loop, as the arithmetic grid and Scan's prefix! already did.

Rel/RelImm emit a BYTE mask. The core's mask idiom is "nonzero is true" and
leaves are width-tagged, so one bit was being stored in eight bytes. Every
consumer now reads any width: Prim::as_byte_mask (borrowed when it already is
one, which is now the common case) for Filter, TryFilter and Select, and
Prim::tags_usize for Branch, TryBranch and Weave -- each dispatching the width
once rather than per element.

And one rule kept the corpus and DDIR working unchanged. A mask meets arithmetic
constantly -- `mul` of two 0/1 masks is an AND, which is how DDIR lowers `And`
and how ten `law` programs combine their conjuncts -- and Bin(Mul, U, 64) on two
byte columns would have stopped typing. So a NARROWER unsigned operand widens to
the cell's declared width: the declared width is the RESULT's. That is the same
sentence as the Reduce rule, applied to the grid. Kind::I and Kind::F encode
order-preservingly at their own width, so they must already match, and say so.

  B1 filter_values 1.680 -> 1.335 ns/row @1M
  B2 cmp_select    2.466 -> 1.778 @1M, 2.203 -> 2.155 @8M

B1 at 8M barely moves: there the Vec<usize> of survivors is 8 bytes per
surviving row and the gather dominates, so the mask is a smaller share. Both are
writer-shaped, the same lever as D5.

Two test expectations changed. No corpus golden moved -- a Vec<u8> and a Vec<u64>
print identically -- and DDIR needs no change.

+++ corgi/src/engine.rs
@@ -40,7 +40,7 @@ mod generators {
 
     /// the mask family: over a list's `bounds` and a per-element 0/1 `mask`, the surviving (nonzero)
     /// positions AND the re-counted per-row bounds, in one pass. Pairs with `gather` to realise `Filter`.
-    pub(crate) fn filter_mask(bounds: &Bounds, mask: &[u64]) -> (Vec<usize>, Vec<usize>) {
+    pub(crate) fn filter_mask(bounds: &Bounds, mask: &[u8]) -> (Vec<usize>, Vec<usize>) {
         let mut idx = Vec::new();
         let mut nb = Vec::with_capacity(bounds.len());
         let mut start = 0;
@@ -260,7 +260,7 @@ pub(crate) fn gather_lanes(srcs: &[Option<&Value>], tags: &[usize], off: &[usize
 /// position) has no constant slot to blend into, so it falls back to the two-source [`gather_lanes`]
 /// — the same split `scatter` makes, and for the same reason. The split is per LEVEL, not per value:
 /// a product blends each leaf field directly and only gathers the fields that need it.
-pub(crate) fn blend(mask: &[u64], then: Value, els: Value) -> Value {
+pub(crate) fn blend(mask: &[u8], then: Value, els: Value) -> Value {
     match (then, els) {
         (Value::Prim(t), Value::Prim(e)) => Value::Prim(t.blend(e, mask)),
         (Value::Prod(ts), Value::Prod(es)) => {
@@ -378,12 +378,12 @@ mod tests {
     #[test]
     fn blend_matches_the_gather_lanes_path() {
         // the general path, written out: row i from lane `mask[i] != 0`, at its own position.
-        fn oracle(mask: &[u64], then: &Value, els: &Value) -> Value {
+        fn oracle(mask: &[u8], then: &Value, els: &Value) -> Value {
             let tags: Vec<usize> = mask.iter().map(|&m| (m != 0) as usize).collect();
             let off: Vec<usize> = (0..tags.len()).collect();
             gather_lanes(&[Some(els), Some(then)], &tags, &off)
         }
-        let mask = [1u64, 0, 0, 1];
+        let mask = [1u8, 0, 0, 1];
         let list = |ends: Vec<usize>, xs: &[u64]| Value::List(ends.into(), Box::new(u(xs)));
 
         // leaf, product of leaves, unit — the lane-wise path.
